Revert "Crucible/All: Deduplication of override matching code (#530)"
This reverts commit 83fb416ab23844a6d96888c660592e8eec331128.
